+++ examples/basic-customize-pixi.html
@@ -1,0 +1,51 @@
+<html lang="en">
+<head>
+	<meta charset="UTF-8">
+	<meta name="viewport" content="width=device-width, initial-scale=1.0">
+	<meta http-equiv="X-UA-Compatible" content="ie=edge">
+	<title>Basic Example - StoryScroll</title>
+	<script src="https://pixijs.download/release/packages/constants.min.js"></script>
+	<script src="https://pixijs.download/release/packages/math.min.js"></script>
+	<script src="https://pixijs.download/release/packages/runner.min.js"></script>
+	<script src="https://pixijs.download/release/packages/settings.min.js"></script>
+	<script src="https://pixijs.download/release/packages/ticker.min.js"></script>
+	<script src="https://pixijs.download/release/packages/utils.min.js"></script>
+	<script src="https://pixijs.download/release/packages/display.min.js"></script>
+	<script src="https://pixijs.download/release/packages/core.min.js"></script>
+	<script src="https://pixijs.download/release/packages/loaders.min.js"></script>
+	<script src="https://pixijs.download/release/packages/sprite.min.js"></script>
+	<script src="https://pixijs.download/release/packages/app.min.js"></script>
+	<script src="https://pixijs.download/release/packages/graphics.min.js"></script>
+	<script src="https://pixijs.download/release/packages/sprite-animated.min.js"></script>
+	<script src="https://pixijs.download/release/packages/spritesheet.min.js"></script>
+	<script src="https://pixijs.download/release/packages/text.min.js"></script>
+	<script src="https://pixijs.download/release/packages/interaction.min.js"></script>
+	<script>
+	PIXI.Renderer.registerPlugin('batch', PIXI.BatchRenderer);
+	PIXI.Renderer.registerPlugin('interaction', PIXI.interaction.InteractionManager);
+	PIXI.Application.registerPlugin(PIXI.AppLoaderPlugin);
+	PIXI.Application.registerPlugin(PIXI.TickerPlugin);
+	PIXI.Loader.registerPlugin(PIXI.SpritesheetLoader);
+	</script>
+	<script src="../lib/storyscroll.min.js"></script>
+</head>
+<body>
+	<script>
+	// init controller
+	var story = new StoryScroll ({
+		crop: 'longside',	// false, none,longside,shortside
+		cropOrigin: 'top',	// center, top,bottom, left,right
+		orientation: 'landscape',
+		scrollDirection: 'x',
+		maxScroll: 20000,
+		backgroundColor: 0x000000
+	});
+	
+	// create a graphic
+	var graphic = story.graphic({x:40, y:540}).act({x: 0,y:400}, 0.8);
+	graphic.beginFill(0xDE3249);
+	graphic.drawRect(50, 50, 100, 100);
+	graphic.endFill();
+	</script>
+</body>
+</html>
