+++ package.json
@@ -1,8 +1,11 @@
 {
   "name": "storyscroll",
-  "version": "2.1.0",
+  "version": "2.2.0",
   "description": "The javascript library for story scroll interactions with html5 canvas",
-  "main": "index.js",
+  "main": "lib/storyscroll.cjs.js",
+  "module": "lib/storyscroll.es.js",
+  "umd": "lib/storyscroll.min.js",
+  "umd:name": "StoryScroll",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "demo": "node ./examples/server-script.js",
@@ -21,20 +24,32 @@
   },
   "homepage": "https://github.com/xxlh/StoryScroll#readme",
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-plugin-transform-es2015-modules-umd": "^6.24.1",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-stage-2": "^6.24.1",
+    "@babel/plugin-proposal-class-properties": "^7.5.5",
+    "@babel/preset-env": "^7.6.0",
     "express": "^4.17.1",
-    "rimraf": "^3.0.0"
+    "gzip-size": "^5.1.1",
+    "pretty-bytes": "^5.3.0",
+    "rimraf": "^3.0.0",
+    "rollup": "^1.21.2",
+    "rollup-plugin-babel": "^4.3.3",
+    "rollup-plugin-buble": "^0.19.8",
+    "rollup-plugin-commonjs": "^10.1.0",
+    "rollup-plugin-node-builtins": "^2.1.2",
+    "rollup-plugin-node-globals": "^1.4.0",
+    "rollup-plugin-node-resolve": "^5.2.0",
+    "uglify-js": "^3.6.0"
   },
   "babel": {
     "presets": [
-      "env",
-      "stage-2"
+      [
+        "@babel/preset-env",
+        {
+          "modules": false
+        }
+      ]
     ],
     "plugins": [
-      "transform-es2015-modules-umd"
+      "@babel/plugin-proposal-class-properties"
     ]
   },
   "dependencies": {
@@ -54,8 +69,10 @@
     "@pixi/text": "^5.1.3",
     "@pixi/ticker": "^5.1.3",
     "@pixi/utils": "^5.1.3",
+    "acorn": "^6.3.0",
     "gsap": "^2.1.3",
     "judgebrowser": "^1.0.0",
-    "pixi-projection": "^0.3.5"
+    "pixi-projection": "^0.3.5",
+    "the-answer": "^1.0.0"
   }
 }
